requirements.txt
Django==5.0.*
django-solo==2.1.*
black==23.12.*
markdown==3.5.*
bleach==6.1.*
django-compressor==4.*
django-csp==3.7
slack_bolt==1.18.*
huey==2.5.*
redis==5.0.*
css-inline==0.13.*